+++ R/imports.R
@@ -1,0 +1,3 @@
+
+#' @importFrom lifecycle deprecated
+NULL