+++ R/onetime.R
@@ -28,6 +28,11 @@
 #' be run, so it may be run repeatedly. Conversely, if the call gives an error,
 #' the lockfile is still written.
 #'
+#' `expiry` is backward-looking. That is, `expiry` is used at check time to see if
+#' the lockfile was written after `Sys.time() - expiry`. It is not used when
+#' the lockfile is created. So, you should set `expiry` to the same value whenever
+#' you call `onetime_do()`. See the example.
+#'
 #' @return The value of `expr`, invisibly; or `default` if `expr` was not run
 #' because it had been run already.
 #'
@@ -41,7 +46,17 @@
 #'   onetime_do(print("printed once"), id = id)
 #' }
 #'
+#' # expiry is "backward-looking":
+#' id2 <- sample(10000L, 1L)
+#' expiry <- as.difftime(1, units = "secs")
+#' onetime_do(print("Expires quickly, right?"), id = id2, expiry = expiry)
+#' Sys.sleep(2)
+#' onetime_do(print("This won't be shown..."), id = id2)
+#' onetime_do(print("... but this will"), id = id2, expiry = expiry)
+#'
+#'
 #' onetime_reset(id = id)
+#' onetime_reset(id = id2)
 #' options(oo)
 onetime_do <- function(
         expr,