+++ R/utils.R
@@ -37,6 +37,42 @@ onetime_reset <- function (
 }
 
 
+#' Mark an action as done
+#'
+#' This manually marks an action as done.
+#'
+#' Note that no `expiry` parameter is available, because `expiry` is
+#' backward-looking. See [onetime_do()] for more information.
+#'
+#' Marking an action done requires permission to store files on the user's
+#' computer, just like other onetime actions.
+#'
+#' @inherit common-params
+#'
+#' @return Invisible `TRUE` if the action represented
+#' by `id` had not been done before, and has now been explicitly marked as done.
+#' Invisible `FALSE` if it was already marked as done (and still is).
+#' @export
+#'
+#' @examples
+#' oo <- options(onetime.dir = tempdir(check = TRUE))
+#' id <- sample(10000L, 1)
+#'
+#' onetime_mark_as_done(id = id)
+#' onetime_message("Won't be shown", id = id)
+#'
+#' onetime_reset(id = id)
+#' options(oo)
+onetime_mark_as_done <- function (
+        id   = calling_package(),
+        path = default_lockfile_dir()
+) {
+  force(id)
+  force(path)
+  invisible(onetime_do(TRUE, id = id, path = path, default = FALSE))
+}
+
+
 #' Check if a onetime call has already been made
 #'
 #' @inherit common-params